[IMP] account_reports/saas~14.4: introduce tax units
If it was configured, remove the old multicompany config parameter that was used by the tax report and log a migration report informing the user that he'll have to configure tax units. We have no way of automating that, as we can't guess how the user was grouping companies when submitting his report.

Goes with:
https://github.com/odoo/odoo/pull/70588
https://github.com/odoo/enterprise/pull/18234

closes odoo/upgrade#2482

Related: odoo/enterprise#18234
Related: odoo/odoo#70588
Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>