[FIX] base@tests: crawling mock test, skip SQL view with custom columns
SQL views can be re-created/re-generated during an upgrade,
therefore overwriting / deleting the extra columns that can
be added by a custom module.

While the fields are still defined in `ir_model_fields`,
and therefore loaded as if they were manual fields thanks to
`util.custom_module_field_as_manual`.

Therefore, when attempting to read the custom field,
whether because of a prefetch or because a custom view
add the field in the report view,
an exception `psycopg2.errors.UndefinedColumn` occurs,
as the custom column/field was removed upon the SQL view regeneration.

It's useful these custom fields in report SQL views are loaded
as if they were manual, as it allows to validate the XML views
for instance.

However, for the crawling mock test,
unfortunately, there is no easy to handle these cases,
which can only happen using custom modules
(this is not possible to add a field in a report SQL view using studio).

Therefore, we choose to skip the crawling of these views for now if
they contain custom fields.

upg-15188

closes odoo/upgrade#2484

Signed-off-by: Christophe Simonis <chs@odoo.com>