[IMP] gmail: allow to upload attachments when logging an email
Purpose
=======
Allow to upload the attachments when logging the email on the partner /
lead / ticket.

Technical
=========
The attachments are base 64 encoded and then added in the JSON posted
on the Odoo endpoint.

The maximum POST request size on the Gmail side is 50 MB. So when the
attachments total size exceed 40 MB we just ignore them and show an
error message.

N.B.: when creating an email from Gmail, if the sum of the file size is
more than 25 MB, they are automatically uploaded on GDrive and the links
to those files are added in the email.

Links
=====
Task 2545048
See odoo/odoo/pull/71372