[REM] (helpdesk),_timesheet@saas~14.5: remove the field and view
The purpose of the task is to 'Assign To Me' button based on
user and remove the timer button on the ticket kanban.

Move the commercial_partner_id field to the Helpdesk module

so in this commit remove the is_self_assigned field and
helpdesk_timer_ticket_view_kanban kanban view.

Related Ent PR: odoo/enterprise#18621

closes odoo/upgrade#2499

Taskid: 2451722
Related: odoo/odoo#72679
Signed-off-by: Christophe Simonis <chs@odoo.com>