[IMP] Outlook: add attachments in logged message
Append attachments to email information when logging an email onto a record

Attachments and images are usually of great importance in an email.
If they are not forwarded, those logged messages thus don't make much sense,
and users have to download each and every attachment, and then go in Odoo and
upload them there, which is the opposite of what we were trying to do with
the plugins.

Task-2545067

COM-PR: https://github.com/odoo/odoo/pull/71543