[IMP] gmail: allow to create Task from the gmail plugin
Purpose
=======
Allow users to connect their tasks to their Gmail mailbox.

Specifications
==============
Show the tasks related to the current partner (the sender of the opened
email) and allow to open them in Odoo.

Allow the users to create a task from the current email. He could
search the projects in which he want to create the task or even create
a new project.

Task 2518668
See odoo/odoo/pull/71542