[IMP] website_event_track_quiz/saas~14.5: removed field 'event_track_ids' and view `event_quiz_view_form_from_track`
Previously, `event_track_ids` field  was used to set `event_track_id`
field on the quiz, but with the linked community PR, the `event_track_id`
will be set by default on creation of the quiz, and so this field is not
required anymore and is removed with this commit.

This commit also removes the view `event_quiz_view_form_from_track` as the
original form view is now utilized everywhere instead of this one.

TaskID-2486552
COM PR odoo/odoo#70067

closes odoo/upgrade#2522

Signed-off-by: Christophe Simonis <chs@odoo.com>