[ADD] base/13.0: handle relative date strings
Due to the new ORM-pocalypse, the values passed to `.write()` now goes
through the cache, which validates the value.
For date and datetime, it only accepts ISO dates.
In previous versions, the value was directly send to PostgreSQL and take
advantage of the coercion of strings to date and timestamps. However,
there are some values that were valid in PostgreSQL, that aren't anymore
in the python version.

opw-2558811

closes odoo/upgrade#2543

Signed-off-by: Christophe Simonis <chs@odoo.com>