[FIX] util: replace_record_references_batch, handle deduplication in mail_followers
When replacing records references,
it can happen the old and the new record were both followers of the
same records, and this raised a constraint
when replacing the old reference to the new reference.

This deduplication case was already handled for regular foreign keys
in m2m, but it wasn't the case for indirect references such as
`mail_followers` (`res_model`, `res_id`).

upg-17289

closes odoo/upgrade#2571

Signed-off-by: Christophe Simonis <chs@odoo.com>