[FIX] gmail: set the user companies when redirecting to Odoo
Bug
===
In a multi-company environment, if the user click on a record (partner,
lead, task...) and if this record does not belong to the current
company, the user will have an access error.

To fix that, we return the companies of the current user and we add
them in the URL, so when the user is redirected to Odoo, all his
companies are selected and no access error occurs.

Task 2542179