[FIX] base@tests: mock crawl, do not rollback custom_module_field_as_manual
When the loading of a menu failed,
the transaction is rollbacked,
to avoid re-using a broken cursor.

But, it was rollbacking the statements done by
`custom_module_field_as_manual`.
For instance, the disabled record rules were
re-enabled because of this rollback.

For upg-17402, this resulted to a custom
record rule to not being disabled in the test prepare phase,
while it was disabled in the test check phase,
and this leaded to a disprancy, with a menu not failing
in the prepare phase while it should have,
and the menu failing in the check phase.

upg-17402

closes odoo/upgrade#2595

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>