[FIX] account: Exclude 0 subtotals when updating grouped account move lines
The goal of _compute_invoice_line_grouped_in_move_line is to update grouped
account move lines with the price information from the original invoice lines,
making the lines visible on the invoice. This is to ensure the subtotals displayed
on the invoice after the migration are correct.

However, earlier on there were new account move lines created for invoice lines
that had a 0 price_subtotal. These got updated by accident when intending to update
grouped account move lines, which would result in all of the involved lines getting
the same subtotal. This would lead to absurd invoices where you could have 4
invoice lines with a subtotal of 1600 each, on an invoice with a total that was
also 1600.

This fix excludes those lines from the update, while still including lines with a
NULL subtotal, since the original grouped account move lines won't have that value
set.

closes odoo/upgrade#2617

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>
Co-authored-by: Denis Ledoux <dle@odoo.com>