[IMP] survey(_*)/saas~14.5: add end_datetime on user input
In order to know average duration of user inputs we now set a datetime when
an answer is closed. At migration we consider that last write on an answer
is probably a good guess to set a default ending datetime.

Also rename a view in enterprise while updating survey views.

Task-2634805
COM PR odoo/odoo#72298
ENT PR odoo/enterprise#19082

closes odoo/upgrade#2624

Signed-off-by: Christophe Simonis <chs@odoo.com>