[IMP] mail_plugin: enrich existing companies on the Gmail plugin
Purpose
=======
Allow users to enrich and update existing Odoo company from the Gmail
plugin so they can have more information on their contact.

Specifications
==============
When the company exist on the Odoo database but is not enriched, show
a button "Enrich" to enrich the company (and write on the fields).

If the current contact is the company itself, hide the company card to
not have redundant information.

Links
=====
Task-2567566
See odoo/odoo/pull/73762
See odoo/mail-client-extensions/pull/15