[FIX] base/9.0.1.3: dont fail early
Invalid views are properly handled (disabled) here:
https://github.com/odoo/upgrade/blob/7579b9fba8a270036f9e6339b490d11f3ae4c505/migrations/base/0.0.0/pre-models-ir_ui_view.py

Moreover, at this stage of the migration (v9.0) it's probably too early
to check the details of how to make the views work again.
For example,
```
Traceback (most recent call last):
  File "/home/odoo/src/odoo/9.0/openerp/service/server.py", line 894, in preload_registries
    registry = RegistryManager.new(dbname, update_module=update_module)
  File "/home/odoo/src/odoo/9.0/openerp/modules/registry.py", line 386, in new
    openerp.modules.load_modules(registry._db, force_demo, status, update_module)
  File "/home/odoo/src/odoo/9.0/openerp/modules/loading.py", line 318, in load_modules
    report=report, models_to_check=models_to_check
  File "/home/odoo/src/odoo/9.0/openerp/modules/loading.py", line 136, in load_module_graph
    migrations.migrate_module(package, 'pre')
  File "/home/odoo/src/odoo/9.0/openerp/modules/migration.py", line 160, in migrate_module
    migrate(self.cr, installed_version)
  File "/home/odoo/src/odoo/9.0/openerp/addons/base/maintenance/migrations/base/9.0.1.3/pre-00-early-fail.py", line 44, in migrate
    raise util.MigrationError('\n'.join(msgs))
MigrationError: Custom view validation failed:
View #313: inheritance may not use attribute 'string' as a selector.
View #314: inheritance may not use attribute 'string' as a selector.
View #325: inheritance may not use attribute 'string' as a selector.
View #316: inheritance may not use attribute 'string' as a selector.
```
shows we need to change the selector, but the right selector is better
chosen later, at the end of the migration (v14.0).

upg-18967

closes odoo/upgrade#2676

Signed-off-by: Christophe Simonis <chs@odoo.com>