[FIX] outlook: fix the pruning of the company cache
Purpose
=======
When a user open a company email, we store the IAP enrich information
in the local storage do not ask IAP each time we open the email.

This cache is automatically cleaned if we exceed a limit by a certain
percentage. But currently, all the cache is cleaned.

Technical
=========
We should keep a certain amount of record in the cache (75% by default),
when we create the "CompanyCache" we give "pruningRatio", this value
should be between 0 and 1 and not between 0 and 100.

Remove the function "getAllcacheObjects" that's not used (and not really
useful as we can access the local storage with the chrome debugger).

Task-2612788