[IMP] outlook: show an error when the database is not reachable
Purpose
=======

Currently the application will display a 404 error message if the mail
plugin module is not installed or if the user made an error in the URL
of his Odoo database. This is confusing for end users.

Instead we now show an error message during the authentication process if
the database is not reachable.

Technical
=========

To not add a new Odoo endpoint, we request one authentication endpoint
which has `cors=*` so it's accessible from Javascript on a different
domain.

Task-2577531
See odoo/odoo/pull/75350
See odoo/mail-client-extensions/pull/17