[IMP] social/saas~14.5: simplify the "description" field on social.media
The description field on a social media currently stores HTML content.
Which feels completely unnecessary since these records are only meant to be
created as data and the content of this field is very simple (one regular
sentence).

This commit simplifies the description by forcing an update on the related
records to store regular text instead.

We also took this opportunity to rename it to "media_description" to ease
greping.

Task-2613240
ENT PR odoo/enterprise#19985

closes odoo/upgrade#2695

Signed-off-by: Christophe Simonis <chs@odoo.com>