[FIX] base/10.saas~17.1.3: adapt fax res.partner view
When sms module is installed, the x_fax field is added inside the div of
the mobile, which makes it invisible (no label) or looked like the mobile
himself (when no mobile filled).

Proposition of adaptation :
move the x_fax field before the next non-inside-a-div field.
+: less code adaptation (than adapting the view regarding the install
of sms which would occur in a separate script where we do not know
the id of the view)
-: appears after the email instead of the mobile, but who uses faxes
   these days...

closes odoo/upgrade#2715

Signed-off-by: Christophe Simonis <chs@odoo.com>