[IMP] helpdesk/saas~15.1: add privacy_visibility
In this commit, we remove the field visibility_member_ids
and add the field privacy_visibility and populate it.

In order to have the same result as before, we need to do the following:
Set privacy_visibility
    - Set it to 'invited_internal' if visibility_member_ids was set
    - 'internal' if visibility_member_ids was not set

Subscribe visibility_member_ids to teams and tickets of the teams.

Enterprise PR - odoo/enterprise#20190
task - 2590082

closes odoo/upgrade#2758

Signed-off-by: Christophe Simonis <chs@odoo.com>