[IMP] hr,hr_payroll/saas~14.5: remove personal documents tab
This commit makes sure to convert old binary fields into attachment in
the chatter and/or documents in the documents app.
The fields are kept but unused because the salary configurator needs
fields to bind to in order to display and request information during the
salary review.

See odoo/odoo#75506
See odoo/enterprise#20420

Task ID: 2628750

closes odoo/upgrade#2769

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>