[FIX] account@saas~12.4: only handle purchase and sale voucher types
`voucher_type` should only be `sale` or `purchase`
https://github.com/odoo/odoo/blob/05b2b700d51746f6d424b77953bdfed3825ad7f4/addons/account_voucher/models/account_voucher.py#L35-L37

In the request database, they are still
`receipt` and `payment` voucher types.

Those were supposed to be converted to `account.payment`
on upgrade to Odoo 9.0 and removed from the table `account_voucher`:
https://github.com/odoo/upgrade/blob/master/migrations/account_voucher/9.0.1.0/pre-05-payments.py#L108

This database has been upgraded using OpenUpgrade,
and most-likely the fact these vouchers are still there
comes from that fact, that it wasn't handled correctly in
OpenUpgrade.
```
\d account_voucher
                            Table "public.account_voucher"
           Column            |            Type             | Collation | Nullable
-----------------------------+-----------------------------+-----------+----------
 ...
 openupgrade_legacy_9_0_type | character varying           |           |
 ...
```

As these types of vouchers are no longer supposed to be there,
and from the fact they are unreachable by default from the interface
(there are domains like `[('voucher_type', '=', 'sale')]` in the window action)
we simply skip these vouchers when converting vouchers to moves.

If the customer wants them, it has to be handled manually by them after
standard upgrade.

This is to prevent the exception
`Some account.voucher.line have no account.move.line`
to happen.

upg-23478

closes odoo/upgrade#1934

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>