[ADD] website_slides{_forum,_survey}/saas~15.1: migration website_slides
Adding a new module, in order to use dashboards inside of
website_slides.

Adding a migration script for wesbite_slides:
We are removing the table `slide_slide_link` and its field `link_ids`
to create a table `slide_slide_resource` that will contain
resources and links for the additional resources.

The new table is the old `slide_slide_resource` on which
we added a `type` selection field and a `link` text field.

See odoo/odoo#75646
See odoo/enterprise#20480

task-2597345

closes odoo/upgrade#2784

Signed-off-by: Christophe Simonis <chs@odoo.com>