[FIX] account@saas~12.4: reconciliation of voucher lines with changed account
The reconciliation of the voucher lines with their move line highly rely
on the `account_id`, especially when there is no product.

Therefore, when we change the `account_id` to create the move line
of a voucher line for draft/cancelled vouchers, because there was a problem
with the account set in those draft/cancelled voucher lines,
we should update the `account_id` on the voucher line as well,
so the reconciliation between the voucher line and the move
that occurs later in the script works properly.

```
Traceback (most recent call last):
  File "/home/odoo/src/odoo/13.0/odoo/service/server.py", line 1192, in preload_registries
    registry = Registry.new(dbname, update_module=update_module)
  File "/home/odoo/src/odoo/13.0/odoo/modules/registry.py", line 86, in new
    odoo.modules.load_modules(registry._db, force_demo, status, update_module)
  File "/home/odoo/src/odoo/13.0/odoo/modules/loading.py", line 442, in load_modules
    migrations.migrate_module(package, 'end')
  File "/home/odoo/src/odoo/13.0/odoo/modules/migration.py", line 180, in migrate_module
    migrate(self.cr, installed_version)
  File "/tmp/tmp7jc85nuh/migrations/account/saas~12.4.1.1/end-09-account-pocalypse.py", line 2039, in migrate
    migrate_voucher_lines(cr)
  File "/tmp/tmp7jc85nuh/migrations/account/saas~12.4.1.1/end-09-account-pocalypse.py", line 1162, in migrate_voucher_lines
    raise util.MigrationError("Some account.voucher.line have no account.move.line")
odoo.upgrade.util.exceptions.MigrationError: Some account.voucher.line have no account.move.line
```

upg-22613

closes odoo/upgrade#2821

Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>