[IMP] base/0.0.0: refactor disabling custom views
Iterate all views by level, starting by top level views.  When a view
fails:
1. Disable all its children, and recheck.
2. If still failing, then
   a. If standard view then restore from fs
   b. For custom views, try to fix its arch
   c. If custom view cannot be fixed then disable it
3. If childless view doesn't fail then activate children one by one and
   recheck. Recursively fix the child view that fails.

We know that standard children _must_ be correct. If not, restoring
their arch _must_ work. Thus we need to activate them first. We activate
the custom children afterwards one by one. The only drawback here is
when a standard view was modified in such a way that it fails with
standard children but it doesn't if we keep the custom views and we
activate them in the order dictated by the priority.

Notes:
* The fix is guarded with an env var ODOO_MIG_TRY_FIX_VIEWS. This
  variable will be true automatically while testing.
* The check is always performed on the root view being checked, this
  means that when fixing a child view the check is always on the parent
  view.
* The logger of ir_ui_view is disabled for the checks to avoid spam in
  the logs.
* Added test for standard view that cannot be anchored due to custom
  removing field
* More than one fix applied can be applied to each view
* When searching view allow parent views of parent models

closes odoo/upgrade#2853

Signed-off-by: Christophe Simonis <chs@odoo.com>