[REF] outlook: run prettier to automatically format the code
Purpose
=======
A prettier and a linter were configured but seem to never be used.
Currently the code doesn't respect any guidelines and is therefor
sometimes hard to read (very long line, missing spaces / too many
spaces).

To fix that, we configured prettier and we run it on the file of the
project.

From now, before committing we must run prettier to keep the code
consistent.

Task-2601837