[FIX] outlook: fix the button "Create And Enrich Company"
Bug
===
If the contact doesn't exist in the database and if you click on the
button "Create And Enrich" nothing happen.

This is because the error returned by Odoo is not displayed on the
plugin and also because the Odoo endpoint is waiting for the ID of the
contact.

This button should be visible only if the contact exists in the
database.

Task-2601837