[REF] outlook: refactor the models part of the outlook add-in
Purpose
=======
This commit clean the code of the models part of the outlook add-in.

The goal is to respect modern JS guidelines (not use var, use const
when possible...) and to make the code more robust.

Company
-------
We do not need to differentiate a company enriched using the ID.

This is redundant with the attribute "EnrichmentStatus", the ID should
only be used to know if the company exist in the Odoo database or not.

Task-2601837