[REF] outlook: merge the components "ProfileCard" and "ContactItem"
Purpose
=======
Currently, the widget ProfileCard is only used in the widget
ContactItem, so it makes no sense to keep 2 different widgets.

This merge will also simplify the code and the HTML structure.

Technical
=========
The `selectable` property on the ContactItem widget is redundant
with the property "onItemClick", since "selectable" is only False
when onItemClick is falsy.

Task-2601837