[IMP] website_slides/saas~15.3: set vote value to -1, 0 and 1
Update the vote value in slide_slide_partner for likes, dislikes and None.
In this commit we limit each negative value to `-1` and each positive value
to `1`; otherwise `0`.

e.g. if a user has a vote equal to "-8" e.g. it will be set to "-1".

See community odoo/odoo#74171

TaskID-2607416

Part-of: odoo/upgrade#2854