[FIX] util/records: fix replace_record_references_batch
`replace_record_references_batch` takes too long when the list of
id_mapping is huge.

Improvements:
1. No need to create a json map if the ids are not changed, the map is
   basically the identity function
2. Create a temp table with the old_ids to avoid the `NOT IN %(old)s`
   wich is very slow for huge id_mapping (>100K entries)
   https://stackoverflow.com/a/17038097/2997574
3. (Minimal gain) pre-compute model_dest id

The upgrade 21828 takes 26h in total, among which 24h are spent on
https://github.com/odoo/upgrade/blob/c646b1c9538962cde76934d320d9bd79327866c4/migrations/documents/saas%7E12.1.1.0/pre-migrate.py
the bulk of time is spent on the line
https://github.com/odoo/upgrade/blob/c646b1c9538962cde76934d320d9bd79327866c4/migrations/documents/saas%7E12.1.1.0/pre-migrate.py#L87

For this upgrade replace_record_references_batch gets called with an
id_mapping of 500K elements, mapping every id to itself.

upg-21828
opw-2647802

closes odoo/upgrade#2865

Signed-off-by: Christophe Simonis <chs@odoo.com>