[FIX] stock_account/saas~12.4.1.1: fix query
The query that migrates the price history runs out of temp space in
Postgres for huge tables.

Example TB:
```
Traceback (most recent call last):
  File "/home/odoo/src/odoo/13.0/odoo/service/server.py", line 1192, in preload_registries
    registry = Registry.new(dbname, update_module=update_module)
  File "/home/odoo/src/odoo/13.0/odoo/modules/registry.py", line 86, in new
    odoo.modules.load_modules(registry._db, force_demo, status, update_module)
  File "/home/odoo/src/odoo/13.0/odoo/modules/loading.py", line 424, in load_modules
    force, status, report, loaded_modules, update_module, models_to_check)
  File "/home/odoo/src/odoo/13.0/odoo/modules/loading.py", line 315, in load_marked_modules
    perform_checks=perform_checks, models_to_check=models_to_check
  File "/home/odoo/src/odoo/13.0/odoo/modules/loading.py", line 230, in load_module_graph
    migrations.migrate_module(package, 'post')
  File "/home/odoo/src/odoo/13.0/odoo/modules/migration.py", line 180, in migrate_module
    migrate(self.cr, installed_version)
  File "/tmp/tmp4rj9wbfs/migrations/stock_account/saas~12.4.1.1/post-10-account-pocalypse.py", line 236, in migrate
    (SUPERUSER_ID,),
  File "/home/odoo/src/odoo/13.0/odoo/sql_db.py", line 173, in wrapper
    return f(self, *args, **kwargs)
  File "/home/odoo/src/odoo/13.0/odoo/sql_db.py", line 250, in execute
    res = self._obj.execute(query, params)
psycopg2.errors.DiskFull: could not write to file "base/pgsql_tmp/pgsql_tmp2748669.979": No space left on device
```

It seems that this query is selecting (CTE) and inserting on the same
table (stock_valuation_layer).
https://stackoverflow.com/a/50835983/2997574

The table sizes at the moment the query is executed are shown below.
Table                 | Size
--------------------- | -------
stock_valuation_layer | 8459586
product_product       | 1577
product_price_history | 499158

The solution we propose is to use an extra table to store the rows to be
inserted on stock_valuation_layer. This query still cannot be run on one
go since we then get the same error. Thus we run it in parallel split by
id range. Once we move back the result we drop the extra table table.
We don't create the extra table as a TEMP table to avoid the issue.
Extra processing time is gained with some pre-computed filters on
ir_property.

upg-31425

closes odoo/upgrade#2867

Signed-off-by: Christophe Simonis <chs@odoo.com>