[REF] outlook: simplify the "CollapseSection" component
Purpose
=======
The logic of the CollapseSection was introduced in each sub-components
which used the CollapseSection (e.g. in the CompanySection).

It makes a lot of code redundant and can be simplified if we move this
logic in the CollapseSection component.

The "collapse logic" will be remove in each sub-components in the
commit which refactor them.

Task-2601837