[REF] outlook: create a new component "Section"
Purpose
=======
The components LeadsSection, TicketsSection and TasksSection are very
redundant. In each component we have almost the same code to
- Create a record (lead, ticket, task)
- Display the error if needed
- Compute the default "collapse" state (collapsed if no record)
- Show the messages
    - When the partner doesn't exist in the Odoo database
    - When there's no records in the Odoo database
    - To show the record list
- Redirect to Odoo after the record has been created

All of this can be done in a single component to remove redundancy,
force consistency between those 3 sections and make easier to change
the behavior (changes in only 1 file instead of making the same changes
in 3 files).

Task-2601837