[FIX] mail/0.0.0: no user mail alias id
When a partner do not have any email, the user login is not an email
and the `alias_name` is null, we set `alias_name` = user login to
create a valid email by adding the `mail.catchall.domain` after the
username.
If the user do not have any mail alias (in the `res_users.alias_id`
column), the `alias_name` can not be updated, then we need to update
it by an other way.

```bash
Traceback (most recent call last):
  File "/home/odoo/src/odoo/14.0/odoo/service/server.py", line 1199, in preload_registries
    registry = Registry.new(dbname, update_module=update_module)
  File "/home/odoo/src/odoo/14.0/odoo/modules/registry.py", line 89, in new
    odoo.modules.load_modules(registry._db, force_demo, status, update_module)
  File "/home/odoo/src/odoo/14.0/odoo/modules/loading.py", line 457, in load_modules
    force, status, report, loaded_modules, update_module, models_to_check)
  File "/home/odoo/src/odoo/14.0/odoo/modules/loading.py", line 349, in load_marked_modules
    perform_checks=perform_checks, models_to_check=models_to_check
  File "/home/odoo/src/odoo/14.0/odoo/modules/loading.py", line 174, in load_module_graph
    migrations.migrate_module(package, 'pre')
  File "/home/odoo/src/odoo/14.0/odoo/modules/migration.py", line 180, in migrate_module
    migrate(self.cr, installed_version)
  File "/tmp/tmpo7cnmbm2/migrations/mail/0.0.0/pre-ensure-admin-email.py", line 40, in migrate
    [username] = cr.fetchone()
TypeError: 'NoneType' object is not iterable
```

upg-41098

closes odoo/upgrade#2908

Signed-off-by: Christophe Simonis <chs@odoo.com>