[IMP] account@10.saas~18: speed-up `price_total` computation
Before this revision, the `price_total` field is computed invoice line
by invoice line, which can take a while when there is a huge number
of invoice lines.

The strategy here is to compute together the lines that at the end
will have the same price total,
when they share the same price, quantity and taxes.
Because there is a high chance when there is a lot of invoice lines
that a lot of them sells the same thing at the same price within the same
quantity.

By default, `compute_all` doesn't depend on the product or the partner,
it's only when the invoice lines uses a python tax that potentially
it could uses the product or the partner, as its passed in the
evaluation context so users could potentially use them.
https://github.com/odoo/odoo/blob/11.0/addons/account/models/account.py#L913-L1020
https://github.com/odoo/odoo/blob/11.0/addons/account/models/account.py#L876-L899
https://github.com/odoo/odoo/blob/11.0/addons/account_tax_python/models/account_tax.py#L39-L47

So, group the invoice line by taxes, price, currency and quantity
when the line doesn't use a python taxes, and compute them together,
otherwise add the product and partner in addition.

This allows to speed up the processing speed from 5 hours to 30 minutes
for the request 31554
2021-09-21 19:26:05,987 28 INFO db_31554 odoo.addons.base.maintenance.migrations.account.saas-18.end-invoice-line-total: [99.83%] 1145820/1147781 invoice lines processed in 5:06:08.807473 (total estimated time: 5:06:40.244550)

upg-31554

closes odoo/upgrade#2915

Signed-off-by: Christophe Simonis <chs@odoo.com>