[FIX] account/tests: make test pass
Restrict the journals used in the payment test to the ones with inbound
payment methods.

When the module `account_check_printing` is installed, the field
`payment_method_id` is recomputed when the partner change, which happens
when we recompute the field `destination_account_id`.
If there are no inbound methods set on the journal, the field
`payment_method_id` is set to `False`, which fails a constraint.

closes odoo/upgrade#2928

Signed-off-by: Christophe Simonis <chs@odoo.com>