[FIX] base/saas~14.5.1.3: adds initial value to column company_details
In task-2355704 a new field was added to res.company. However, it needs
to be initialized with the correct data in order to render the pdf
correctly. This PR calls the default method to set this value during the
end phase of the base upgrade.

closes odoo/upgrade#2946

Related: https://github.com/odoo/odoo/pull/78652
Signed-off-by: Olivier Dony (odo) <odo@openerp.com>