[FIX] theme_common@0.0.0: reset arch_updated when copy is untouched
In standard odoo/odoo, there is a mechanism to update the copies
of a theme template (`theme.ir.ui.view`) when the template gets updated
https://github.com/odoo/odoo/blob/bbdc3c4d09433a1e42318133cd12d8ffa33dc049/addons/website_theme_install/models/ir_module_module.py#L111

It relies on the fact the copy has been left untouched to update with
the source of the template, thanks to a `arch_updated` flag:
https://github.com/odoo/odoo/blob/bbdc3c4d09433a1e42318133cd12d8ffa33dc049/addons/website_theme_install/models/ir_module_module.py#L165

However, due to a bug in the past, the flag `arch_updated` could be set to `True`
during a module update, therefore marking the view as modified by the user,
while it was not, and therefore preventing this mechanism to update the copies
when a theme get updated during an upgrade (from 13.0 to 14.0 for instance)

```sql
dle_42172=> SELECT count(*), arch_updated FROM ir_ui_view WHERE theme_template_id IS NOT NULL GROUP BY arch_updated;
 count | arch_updated
-------+--------------
   172 | t
```

In the case of the studied request,
this prevented the view `theme_common.assets_frontend` to be updated during the upgrade from 13.0 to 14.0
and to remove
`/theme_common/static/src/scss/s_google_map.scss`
`/theme_common/static/src/js/s_google_map_frontend.js`
from the asset view,
therefore leading to the raise of an exception when loading the website upgraded in 14.0.

This revision aims to reset this `arch_updated` to `False` when the copy arch is identical to the template arch
before doing the upgrade.

upg-42172

closes odoo/upgrade#2948

Signed-off-by: Christophe Simonis <chs@odoo.com>