[FIX] point_of_sale/saas~11.5.1.0.1: fix query
Disabling FK triggers is not possible without super_user rights.
OTOH the reason this was in place was to avoid multiple updates of the
same record on the same transaction, which could be avoided with a
commit. This is a side effect of parallel execute, thus using that here.

upg-43496

closes odoo/upgrade#2962

Signed-off-by: Christophe Simonis <chs@odoo.com>