[IMP] appointment/saas~15.3: add answer input model for reporting
Previously, answers and questions were both many2many fields
on each other's models. As we do not want answers to be
suggested to other appointments / other questions, we change
the answer_ids to a one2many field on the question model,
and the question_ids to a question_id many2one field on the
answer model.

The script duplicates answers that were previously shared
to make them unique per question_id. The previous relational
table is removed. It also duplicates corresponding translations
if any.

A sequence field is also added to the answer model, to make
answers orderable.

---   LINKS   ---
Task-2621327
ENT-PR odoo/enterprise#20433

closes odoo/upgrade#2984

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>