[FIX] pos_loyalty/saas~15.1 : delete rules and rewards without program
Steps :

Create a new POS and activate option Loyalty program.
Go to POS loyalty programs and create a new one.
Create a rule and/or a reward with a new product.
Delete the program.
Try to archive the product.

Issue :

Deleting loyalty programs does not delete loyalty rules not loyalty rewards.
As a consequence, we get an error message when we then try to archive a product.

Solution :

Add option "ondelete='cascade'" on both rules and rewards to delete them
when deleting program.

Migration for PR: https://github.com/odoo/enterprise/pull/21857

opw-2662333

closes odoo/upgrade#2991

Related: odoo/enterprise#21857
Signed-off-by: Christophe Simonis <chs@odoo.com>