[FIX] l10n_be_reports/15.0.1.1: move view to bridge module
Currently, the view `res_partner_view_form_inherit_sms`
is located in `l10n_be_reports`.

The issue is that this view inherits the module `sms`,
which is not a dependency of `l10n_be_reports`.

This is why the view has been moved to a dedicated
bridge module.

Related PR : odoo/enterprise#21982

closes odoo/upgrade#3012

Signed-off-by: Christophe Simonis <chs@odoo.com>