[FIX] account@0.0.0: do not unassign custom tags from accounts
Related to odoo/odoo#79602

Basically, in `l10n_at` (at least) the list of tags on the `account.account.template`
is set with a `[(6, )]` command.
In other words, the list is reset at every `-u` on the module,
discarding custom tags added on the template.

For instance, Wolfgang adds custom tags on these account templates
with a custom module `l10n_at_account_tag_hierarchy`,
adding them with a `[(4,)]` command.

Because the list of tags is reset during the `-u`,
the custom tags were being removed from the accounts as well
with that script, while they should remain.

upg-45351

closes odoo/upgrade#3021

Signed-off-by: Christophe Simonis <chs@odoo.com>