[ADD] account/9.saas~13.1:account.reconcile seq
This commit aims to gives a default sequence to account.reconcile as it
would be for any brand new database.

context:
In previous version of Odoo, the seq for account.reconcile was company
specific.
It wasn't a problem because client were leveraging a bug that was
corrected on this commit: 76824f53a29d625b0569410aec0efd1bea994c40

After this, a company without sequence for account.reconcile couldn't
reconcile any account.move.line.

Ticket on which the issue was found: #2666178

closes odoo/upgrade#3022

Signed-off-by: Brice Bartoletti <bib@odoo.com>
Co-authored-by: Christophe Simonis <KangOl@users.noreply.github.com>
Co-authored-by: SEINLET Nicolas <nse@odoo.com>