[FIX] sale_coupon_taxcloud/saas~12.4.1.1: create columns
During the account pocalypse the column price_taxcloud should be moved
from account_invoice_line to account_move_line.
Here we do as in `migrations/inter_company_rules/saas~12.4.1.1`:
we create the column them copy the values using invl_aml_mapping.
https://github.com/odoo/enterprise/blob/ef3eb0e05e216c737994b0e0a4822eed9511faa6/sale_coupon_taxcloud/models/account_invoice.py#L23

The ORM takes a lot of time creating the column, and it fills it with 0.
```
2021-11-02 18:33:14,723 2229 INFO db_42897 odoo.modules.registry: module sale_coupon_taxcloud: creating or updating database tables
2021-11-02 19:26:40,037 2229 INFO db_42897 odoo.addons.base.models.ir_translation: module sale_coupon_taxcloud: loading base translation file fr for language fr_BE
```

upg-42897
opw-2672031

closes odoo/upgrade#3053

Signed-off-by: Christophe Simonis <chs@odoo.com>