[FIX] project_account/saas~15.1: module for analytic accounting setting
Before this fix, the "Track costs & revenues by project" option is missing
from the project module if both project and invoicing are installed.

Now, this setting is shown on the project module until accounting is installed,
as expected. It is then shown in the settings of the account module.

task-2695812

See odoo/odoo#80045
See odoo/enterprise#22370

closes odoo/upgrade#3051

Signed-off-by: Christophe Simonis <chs@odoo.com>