[FIX] sales_team/saas~14.2.1.1: fix permission error in crawler
When crm_team_member points to an inactive user, there may be permission
errors later when crawling `Sales > Orders > Sales Teams`.

Example TB:
```
 Traceback (most recent call last):
   File "/home/odoo/src/odoo/15.0/odoo/fields.py", line 1083, in __get__
    recs._fetch_field(self)
   File "/home/odoo/src/odoo/15.0/odoo/models.py", line 3276, in _fetch_field
    self._read(fnames)
   File "/home/odoo/src/odoo/15.0/odoo/addons/base/models/res_users.py", line 449, in _read
    super(Users, self)._read(fields)
   File "/home/odoo/src/odoo/15.0/odoo/models.py", line 3387, in _read
    raise self.env['ir.rule']._make_access_error('read', forbidden)
 odoo.exceptions.AccessError: Due to security restrictions, you are not allowed to access 'Users' (res.users) records.

Records: ... (id=2)
User: ... (id=6)

This restriction is due to the following rules:
- user rule

Note: this might be a multi-company issue.
```
The user 6 (current admin) is not able to open the crm teams because the
user 2 (former admin, and pointed by `base.admin_user`) is disabled.
```
54654_afu=> select id,active,company_id from res_users where id in (2, 6)
+------+----------+--------------+
| id   | active   | company_id   |
|------+----------+--------------|
| 2    | False    | 3            |
| 6    | True     | 2            |
+------+----------+--------------+
```

This solves
upg-54654
upg-43955
but there are more in the TBG on the server(11) that could _potentially_ be
fixed with this.

closes odoo/upgrade#3096

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>