[FIX] sale/saas~14.4.1.2: block upgrade if UoM mismatch
When the product associated to a sale order line is not the same as the
one of its packaging, as expected, we may get a mismatch on the UoM
categories between the products.

Here we highlight those mismatches, only when the UoM category is not
correct to avoid blocking too many upgrades that currently pass.

Example TB:
```
  File "/tmp/tmphukk6s2z/migrations/sale/saas~14.4.1.2/pre-migrate.py", line 34, in migrate
    qty = float_round(l_uom._compute_quantity(l_qty, p_uom) / p_qty, p_uom.rounding)
  File "/home/odoo/src/odoo/15.0/addons/uom/models/uom_uom.py", line 217, in _compute_quantity
    raise UserError(_('The unit of measure %s defined on the order line doesn\'t belong to the same category as the unit of measure %s defined on the product. Please correct the unit of measure defined on the order line or on the product, they should belong to the same category.') % (self.name, to_unit.name))
odoo.exceptions.UserError: The unit of measure Unit defined on the order line doesn't belong to the same category as the unit of measure kg defined on the product. Please correct the unit of measure defined on the order line or on the product, they should belong to the same category.
```

In 15.0 this check _should_ prevent the product mismatch to happen
https://github.com/odoo/odoo/blob/10732e077d4bfc296ca59a9c3df613b9fd1cb275/addons/sale/models/sale_order_line.py#L313
while in 14.0 there was no such check for equality of the products
https://github.com/odoo/odoo/blob/00b5cf2d90039db3277b866fd146c82a882604a7/addons/sale_stock/models/sale_order.py#L270

upg-53174
upg-46680
potentially 10 in total as per TBG

closes odoo/upgrade#3102

Signed-off-by: Nicolas Seinlet (nse) <nse@odoo.com>